+++ telegram/src/main/java/ru/kampaii/telegram/actions/updates/impl/LeaveGroupUpdateExecutor.java
@@ -40,7 +40,7 @@ public class LeaveGroupUpdateExecutor extends NonCommandUpdateExecutor {
 
     @Override
     public void execute(Update update) {
-        Integer user = update.getMessage().getFrom().getId();
+        Long user = update.getMessage().getFrom().getId();
 
         Long chatId = update.getMessage().getChat().getId();
